+++ sydtest/src/Test/Syd/Output.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE CPP #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiWayIf #-}
 {-# LANGUAGE NumericUnderscores #-}
@@ -23,40 +22,36 @@ import qualified Data.Text as T
 import GHC.Stack
 import Safe
 import Test.QuickCheck.IO ()
+import Test.Syd.OptParse
 import Test.Syd.Run
 import Test.Syd.SpecDef
 import Test.Syd.SpecForest
 import Text.Colour
 import Text.Printf
 
-#ifdef mingw32_HOST_OS
-import System.Console.ANSI (hSupportsANSIColor)
-import System.IO (stdout)
-#else
-import Text.Colour.Capabilities.FromEnv
-#endif
+printOutputSpecForest :: Settings -> Timed ResultForest -> IO ()
+printOutputSpecForest settings results = do
+  tc <- deriveTerminalCapababilities settings
 
-printOutputSpecForest :: TerminalCapabilities -> Timed ResultForest -> IO ()
-printOutputSpecForest tc results = do
-  forM_ (outputResultReport results) $ \chunks -> do
+  forM_ (outputResultReport settings results) $ \chunks -> do
     putChunksWith tc chunks
     SB8.putStrLn ""
 
-renderResultReport :: TerminalCapabilities -> Timed ResultForest -> Builder
-renderResultReport tc rf =
+renderResultReport :: Settings -> TerminalCapabilities -> Timed ResultForest -> Builder
+renderResultReport settings tc rf =
   mconcat $
     L.intersperse (SBB.char7 '\n') $
-      map (renderChunks tc) (outputResultReport rf)
+      map (renderChunks tc) (outputResultReport settings rf)
 
-outputResultReport :: Timed ResultForest -> [[Chunk]]
-outputResultReport trf@(Timed rf _) =
+outputResultReport :: Settings -> Timed ResultForest -> [[Chunk]]
+outputResultReport settings trf@(Timed rf _) =
   concat
     [ outputTestsHeader,
       outputSpecForest 0 (resultForestWidth rf) rf,
       [ [chunk ""],
         [chunk ""]
       ],
-      outputFailuresWithHeading rf,
+      outputFailuresWithHeading settings rf,
       [[chunk ""]],
       outputStats (computeTestSuiteStats <$> trf),
       [[chunk ""]]
@@ -65,13 +60,13 @@ outputResultReport trf@(Timed rf _) =
 outputFailuresHeader :: [[Chunk]]
 outputFailuresHeader = outputHeader "Failures:"
 
-outputFailuresWithHeading :: ResultForest -> [[Chunk]]
-outputFailuresWithHeading rf =
-  if any testFailed (flattenSpecForest rf)
+outputFailuresWithHeading :: Settings -> ResultForest -> [[Chunk]]
+outputFailuresWithHeading settings rf =
+  if shouldExitFail settings rf
     then
       concat
         [ outputFailuresHeader,
-          outputFailures rf
+          outputFailures settings rf
         ]
     else []
 
@@ -333,12 +328,9 @@ spacingChunk level descriptionText executionTimeText treeWidth = chunk $ T.pack 
           actualMaxWidth = max totalNecessaryWidth preferredMaxWidth
        in actualMaxWidth - paddingSize * level - actualTimingWidth - actualDescriptionWidth
 
-testFailed :: (a, TDef (Timed TestRunResult)) -> Bool
-testFailed = (== TestFailed) . testRunResultStatus . timedValue . testDefVal . snd
-
-outputFailures :: ResultForest -> [[Chunk]]
-outputFailures rf =
-  let failures = filter testFailed $ flattenSpecForest rf
+outputFailures :: Settings -> ResultForest -> [[Chunk]]
+outputFailures settings rf =
+  let failures = filter (testFailed settings . timedValue . testDefVal . snd) $ flattenSpecForest rf
       nbDigitsInFailureCount :: Int
       nbDigitsInFailureCount = floor (logBase 10 (L.genericLength failures) :: Double)
       padFailureDetails = (chunk (T.pack (replicate (nbDigitsInFailureCount + 4) ' ')) :)
@@ -367,6 +359,7 @@ outputFailures rf =
                         chunk $ T.intercalate "." ts
                       ]
                   ],
+                  map padFailureDetails $ retriesChunks testRunResultStatus testRunResultRetries testRunResultFlakinessMessage,
                   map (padFailureDetails . (: []) . chunk . T.pack) $
                     case (testRunResultNumTests, testRunResultNumShrinks) of
                       (Nothing, _) -> []
@@ -535,15 +528,3 @@ orange = colour256 166
 
 darkRed :: Colour
 darkRed = colour256 160
-
-#ifdef mingw32_HOST_OS
-detectTerminalCapabilities :: IO TerminalCapabilities
-detectTerminalCapabilities = do
-  supports <- hSupportsANSIColor stdout
-  if supports
-    then pure With8BitColours
-    else pure WithoutColours
-#else
-detectTerminalCapabilities :: IO TerminalCapabilities
-detectTerminalCapabilities = getTerminalCapabilitiesFromEnv
-#endif