+++ sydtest/src/Test/Syd/Runner.hs
@@ -29,24 +29,21 @@ import Text.Colour
 import Text.Printf
 
 sydTestResult :: Settings -> TestDefM '[] () r -> IO (Timed ResultForest)
-sydTestResult sets spec = do
-  let totalIterations = case settingIterations sets of
+sydTestResult settings spec = do
+  let totalIterations = case settingIterations settings of
         OneIteration -> Just 1
         Iterations i -> Just i
         Continuous -> Nothing
   case totalIterations of
-    Just 1 -> sydTestOnce sets spec
-    _ -> sydTestIterations totalIterations sets spec
+    Just 1 -> sydTestOnce settings spec
+    _ -> sydTestIterations totalIterations settings spec
 
 sydTestOnce :: Settings -> TestDefM '[] () r -> IO (Timed ResultForest)
-sydTestOnce sets spec = do
-  specForest <- execTestDefM sets spec
-  tc <- case settingColour sets of
-    Just False -> pure WithoutColours
-    Just True -> pure With8BitColours
-    Nothing -> detectTerminalCapabilities
-  withArgs [] $ case settingThreads sets of
-    Synchronous -> runSpecForestInterleavedWithOutputSynchronously tc (settingFailFast sets) specForest
+sydTestOnce settings spec = do
+  specForest <- execTestDefM settings spec
+  tc <- deriveTerminalCapababilities settings
+  withArgs [] $ case settingThreads settings of
+    Synchronous -> runSpecForestInterleavedWithOutputSynchronously settings specForest
     ByCapabilities -> do
       i <- fromIntegral <$> getNumCapabilities
 
@@ -64,26 +61,26 @@ sydTestOnce sets spec = do
             chunk "         -threaded -rtsopts -with-rtsopts=-N",
             chunk "         (This is important for correctness as well as speed, as a parallell test suite can find thread safety problems.)"
           ]
-      runSpecForestInterleavedWithOutputAsynchronously tc (settingFailFast sets) i specForest
+      runSpecForestInterleavedWithOutputAsynchronously settings i specForest
     Asynchronous i ->
-      runSpecForestInterleavedWithOutputAsynchronously tc (settingFailFast sets) i specForest
+      runSpecForestInterleavedWithOutputAsynchronously settings i specForest
 
 sydTestIterations :: Maybe Word -> Settings -> TestDefM '[] () r -> IO (Timed ResultForest)
-sydTestIterations totalIterations sets spec =
+sydTestIterations totalIterations settings spec =
   withArgs [] $ do
     nbCapabilities <- fromIntegral <$> getNumCapabilities
 
-    let runOnce sets_ = do
-          specForest <- execTestDefM sets_ spec
-          r <- timeItT $ case settingThreads sets_ of
-            Synchronous -> runSpecForestSynchronously (settingFailFast sets_) specForest
-            ByCapabilities -> runSpecForestAsynchronously (settingFailFast sets_) nbCapabilities specForest
-            Asynchronous i -> runSpecForestAsynchronously (settingFailFast sets_) i specForest
+    let runOnce settings_ = do
+          specForest <- execTestDefM settings_ spec
+          r <- timeItT $ case settingThreads settings_ of
+            Synchronous -> runSpecForestSynchronously settings_ specForest
+            ByCapabilities -> runSpecForestAsynchronously settings_ nbCapabilities specForest
+            Asynchronous i -> runSpecForestAsynchronously settings_ i specForest
           performGC -- Just to be sure that nothing dangerous is lurking around in memory anywhere
           pure r
 
     let go iteration = do
-          newSeedSetting <- case settingSeed sets of
+          newSeedSetting <- case settingSeed settings of
             FixedSeed seed -> do
               let newSeed = seed + fromIntegral iteration
               putStrLn $ printf "Running iteration: %4d with seed %4d" iteration newSeed
@@ -91,8 +88,8 @@ sydTestIterations totalIterations sets spec =
             RandomSeed -> do
               putStrLn $ printf "Running iteration: %4d with random seeds" iteration
               pure RandomSeed
-          rf <- runOnce $ sets {settingSeed = newSeedSetting}
-          if shouldExitFail sets (timedValue rf)
+          rf <- runOnce $ settings {settingSeed = newSeedSetting}
+          if shouldExitFail settings (timedValue rf)
             then pure rf
             else case totalIterations of
               Nothing -> go $ succ iteration
@@ -101,9 +98,5 @@ sydTestIterations totalIterations sets spec =
                 | otherwise -> go $ succ iteration
 
     rf <- go 0
-    tc <- case settingColour sets of
-      Just False -> pure WithoutColours
-      Just True -> pure With8BitColours
-      Nothing -> detectTerminalCapabilities
-    printOutputSpecForest tc rf
+    printOutputSpecForest settings rf
     pure rf